+++ src/main/java/com/maxmind/minfraud/request/Email.java
@@ -412,6 +412,7 @@ public final class Email extends AbstractModel {
         domain = IDN.toASCII(domain);
 
         domain = domain.replaceAll("(?:\\.com){2,}$", ".com");
+        domain = domain.replaceAll("\\.com[^.]+$", ".com");
 
         if (typoDomains.containsKey(domain)) {
             domain = typoDomains.get(domain);