+++ src/test/java/com/maxmind/minfraud/request/EmailTest.java
@@ -199,6 +199,10 @@ public class EmailTest {
         e = new Builder(false).address("foo@example.com.com").hashAddress().build();
         assertEquals(toMD5("foo@example.com"), e.getAddress(), "MD5");
         assertEquals("example.com.com", e.getDomain(), "domain");
+
+        e = new Builder(false).address("foo@example.comfoo").hashAddress().build();
+        assertEquals(toMD5("foo@example.com"), e.getAddress(), "MD5");
+        assertEquals("example.comfoo", e.getDomain(), "domain");
     }
 
     private String toMD5(String s) throws NoSuchAlgorithmException {