+++ CHANGELOG.md
@@ -20,6 +20,9 @@ CHANGELOG
 * Duplicate `.com`s are now removed from email domain names when
   `hashAddress` is used. For example, `example.com.com` will become
   `example.com`.
+* Extraneous characters after `.com` are now removed from email domain
+  names when `hashAddress` is used. For example, `example.comfoo` will
+  become `example.com`.
 
 3.3.0 (2023-12-05)
 ------------------